修改nacos

+++ src/main/java/com/github/novicezk/midjourney/controller/AccountsUpdateController.java
@@ -2,6 +2,7 @@ package com.github.novicezk.midjourney.controller;
 
 import com.alibaba.nacos.api.exception.NacosException;
 import com.fasterxml.jackson.core.JsonProcessingException;
+import com.github.novicezk.midjourney.ProxyProperties;
 import com.github.novicezk.midjourney.domain.DiscordAccount;
 import com.github.novicezk.midjourney.dto.AccountDTO;
 import com.github.novicezk.midjourney.dto.AccountRespondDTO;
@@ -30,51 +31,56 @@ public class AccountsUpdateController {
     private final AccountsRepeatUtils accountsRepeatUtils;
 
     private final DiscordLoadBalancer loadBalancer;
+    private final ProxyProperties properties;
 
     @ApiOperation(value = "修改账号")
-    @PostMapping ("/accounts")
+    @PostMapping("/accounts")
     public Result<String> updateAccounts(@RequestBody List<AccountDTO.Account> list) {
-        //修改本地的yml的accounts
+        //修改本地的yml的account
 //      accountsUpdateUtils.updateConfig(list);
-      //发送新的accounts到nacos
+        //发送新的accounts到nacos
         accountsUpdateUtils.sendAccountsTONacos(list);
-        accountsRepeatUtils.fun(false,null);
-      log.info("账号更新为："+list.toString());
-      return Result.ok("账号更新成功");
+        accountsRepeatUtils.fun(false, null);
+        log.info("账号更新为：" + list.toString());
+        return Result.ok("账号更新成功");
     }
 
     @ApiOperation(value = "获取账号")
-    @GetMapping ("/getInfo")
-    public Map<String,Object> getInfo() {
-       List<AccountRespondDTO> listmap= new ArrayList<>();
-        List<DiscordAccount> list =this.loadBalancer.getAllInstances().stream().map(DiscordInstance::account).toList();
+    @GetMapping("/getInfo")
+    public Map<String, Object> getInfo() {
+        List<AccountRespondDTO> listmap = new ArrayList<>();
+        List<DiscordAccount> list = this.loadBalancer.getAllInstances().stream().map(DiscordInstance::account).toList();
         for (DiscordAccount discordAccount : list) {
             AccountRespondDTO account = new AccountRespondDTO();
-           account.setChannelId(discordAccount.getChannelId());
-           account.setCoreSize(discordAccount.getCoreSize());
-           account.setGuildId(discordAccount.getGuildId());
-           account.setUserToken(discordAccount.getUserToken());
-           account.setQueueSize(discordAccount.getQueueSize());
-           account.setId(discordAccount.getId());
-           account.setEnable(discordAccount.isEnable());
-           account.setTimeoutMinutes(discordAccount.getTimeoutMinutes());
-           listmap.add(account);
+            account.setChannelId(discordAccount.getChannelId());
+            account.setCoreSize(discordAccount.getCoreSize());
+            account.setGuildId(discordAccount.getGuildId());
+            account.setUserToken(discordAccount.getUserToken());
+            account.setQueueSize(discordAccount.getQueueSize());
+            account.setId(discordAccount.getId());
+            account.setEnable(discordAccount.isEnable());
+            account.setTimeoutMinutes(discordAccount.getTimeoutMinutes());
+            listmap.add(account);
         }
-        Map<String,Object> map =new HashMap<>();
-        map.put("total",listmap.size());
-        map.put("row",listmap);
-        map.put("code",200);
-        map.put("msg","查询成功");
+        Map<String, Object> map = new HashMap<>();
+        map.put("total", listmap.size());
+        map.put("row", listmap);
+        map.put("code", 200);
+        map.put("msg", "查询成功");
         return map;
     }
 
     @ApiOperation(value = "删除账号")
     @GetMapping("/delete")
     public Result<String> deleteByGuiId(@RequestParam String guildId) throws JsonProcessingException, NacosException {
-        if(!guildId.isEmpty()){
-            accountsUpdateUtils.deleteByGuiId(guildId);
-            accountsRepeatUtils.fun(true,guildId);
-            return Result.ok("删除账号:"+guildId);
+        if (!guildId.isEmpty()) {
+            String result = accountsUpdateUtils.deleteByGuiId(guildId);
+            if (result.equals("删除成功")) {
+                accountsRepeatUtils.fun(true, guildId);
+                return Result.ok("删除账号:" + guildId);
+            } else {
+                return Result.fail(result);
+            }
         }
         return Result.fail("输入正确的guildId");
     }
@@ -82,20 +88,29 @@ public class AccountsUpdateController {
     @ApiOperation(value = "添加账号")
     @PostMapping("/add")
     public Result<String> addAccount(@RequestBody AccountDTO.Account account) throws JsonProcessingException, NacosException {
-      try{
-          accountsUpdateUtils.addAccount(account);
-          boolean enable= accountsRepeatUtils.fun(false,null);
-          if(enable){
-              return Result.ok("添加账号成功");
-          }else {
-              //把已经添加到本地的删除
-              accountsUpdateUtils.deleteByGuiId(account.getGuildId());
-              Result.fail("添加失败：该账号不可用");
-          }
-      }catch (Exception e){
-          accountsUpdateUtils.deleteByGuiId(account.getGuildId());
-        e.printStackTrace();
-      }
+        if (account.getGuildId() == null) {
+            return Result.fail("添加失败：该账号不能为空");
+        }
+        try {
+            accountsUpdateUtils.addAccount(account);
+            boolean enable = accountsRepeatUtils.fun(false, null);
+            if (enable) {
+                return Result.ok("添加账号成功");
+            } else {
+                //把已经添加到本地的删除
+                accountsUpdateUtils.deleteByGuiId(account.getGuildId());
+                return Result.fail("添加失败：该账号不可用/账号list为0");
+            }
+        } catch (Exception e) {
+            accountsUpdateUtils.deleteByGuiId(account.getGuildId());
+            e.printStackTrace();
+        }
         return Result.fail("添加失败：该账号不可用");
+    }
+
+    @ApiOperation(value = "test账号")
+    @GetMapping("/test")
+    public String deleteByGuiId() throws JsonProcessingException, NacosException {
+    return properties.getAccounts().toString();
     }
 }
